refactor: exception handler 하나로 합치기
