+++ src/main/java/com/spring/board/springboard/domain/GlobalExceptionHandler.java
@@ -1,5 +1,10 @@
 package com.spring.board.springboard.domain;
 
+import com.spring.board.springboard.post.exception.NoPostException;
+import com.spring.board.springboard.user.exception.AuthenticateException;
+import com.spring.board.springboard.user.exception.NoMemberException;
+import com.spring.board.springboard.user.exception.SessionCreateException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -31,6 +36,36 @@ public class GlobalExceptionHandler {
         ErrorResponse errorResponse = new ErrorResponse(illegalArgumentException.getMessage());
 
         return ResponseEntity.badRequest()
+                .body(errorResponse);
+    }
+
+    @ExceptionHandler(NoPostException.class)
+    public ResponseEntity<ErrorResponse> handleNoPostException(NoPostException noPostException) {
+        ErrorResponse errorResponse = new ErrorResponse(noPostException.getMessage());
+
+        return ResponseEntity.status(
+                        HttpStatus.NOT_FOUND)
+                .body(errorResponse);
+    }
+
+    @ExceptionHandler(NoMemberException.class)
+    public ResponseEntity<ErrorResponse> handleNoMemberException(NoMemberException noMemberException) {
+        ErrorResponse errorResponse = new ErrorResponse(noMemberException.getMessage());
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                .body(errorResponse);
+    }
+
+    @ExceptionHandler(AuthenticateException.class)
+    public ResponseEntity<ErrorResponse> handleAuthenticateException(AuthenticateException authenticateException) {
+        ErrorResponse errorResponse = new ErrorResponse(authenticateException.getMessage());
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(errorResponse);
+    }
+
+    @ExceptionHandler(SessionCreateException.class)
+    public ResponseEntity<ErrorResponse> handleSessionCreateException(SessionCreateException sessionCreateException) {
+        ErrorResponse errorResponse = new ErrorResponse(sessionCreateException.getMessage());
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                 .body(errorResponse);
     }
 }
